+++ Libraries/PeriphDrivers/Source/TMR/tmr_common.h
@@ -37,8 +37,8 @@
 #include "tmr.h"
 
 /* **** Functions **** */
-void MXC_TMR_Common_Delay(mxc_tmr_regs_t *tmr, unsigned long us);
-void MXC_TMR_Common_TO_Start(mxc_tmr_regs_t *tmr, unsigned long us);
+void MXC_TMR_Common_Delay(mxc_tmr_regs_t *tmr, uint32_t us);
+void MXC_TMR_Common_TO_Start(mxc_tmr_regs_t *tmr, uint32_t us);
 int MXC_TMR_Common_TO_Check(mxc_tmr_regs_t *tmr);
 void MXC_TMR_Common_TO_Stop(mxc_tmr_regs_t *tmr);
 void MXC_TMR_Common_TO_Clear(mxc_tmr_regs_t *tmr);
